requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
requests==2.25.1
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==1.0.1
uwsgi
psycopg2